Organize iot tests under iot/
