+++ tests/test_plug.py
@@ -1,9 +1,9 @@
 import pytest
 
 from kasa import DeviceType
+from tests.iot.test_iotdevice import SYSINFO_SCHEMA
 
 from .conftest import plug, plug_iot, plug_smart, switch_smart, wallswitch_iot
-from .test_iotdevice import SYSINFO_SCHEMA
 
 # these schemas should go to the mainlib as
 # they can be useful when adding support for new features/devices
